Add support for Django 5.0 by removing deprecated features
There is only one Django feature used here that is removed in >=5.0 which is the `length_is` filter query:

```
RemovedInDjango51Warning: The length_is template filter is deprecated in favor of the length template filter and the == operator within an {% if %} tag.
```

A trivial change to make and the `length` feature if backwards compatile to the earliest support version this repo supports of 2.2.