Fix displaying the installed version of Django-Jazzmin in the UI.
Since we changed our CD process and now will automatically generate test
builds (#550), the version in the `__init__` file now needs to be
updated to stop it from being hardcoded. As in development the version
is now set to 0.0.0 in the `pyproject.toml` file, we need to get the
version dyanmically. We can call the package manager to check which
version of `django-jazzmin` was installed. This sadly does vary
depending on which version of Python is installed - support for these
older versions are soon to be dropped.