Fix fieldsets to follow Bootstrap usage recommendations
Currently Jazzmin inserts all fields in a Django "field group" into a single Bootstrap "field-group" div. However Django field groups and the Bootstrap concept of a field group are not the same thing, meaning that forms with multiple fields in a row break the layout.

A bootstrap field group should only contain a single label and input pair, and layout of the fields should be controlled with the row and col classes (see https://getbootstrap.com/docs/4.6/components/forms/). My edits follow this design pattern, and place each Django field group in a row, with each individual field in a form-group that is also a col.